:sparkles: 동시 요청 처리를 위해 handler 생성
