+++ src/main/java/was/HttpServer.java
@@ -1,16 +1,15 @@
 package was;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static util.MyLog.log;
 
 public class HttpServer {
+    private final ExecutorService executor = Executors.newFixedThreadPool(10);
     private final int port;
 
     public HttpServer(int port) {
@@ -23,71 +22,7 @@ public class HttpServer {
 
         while (true) {
             Socket socket = serverSocket.accept();
-            process(socket);
-        }
-    }
-
-    private void process(Socket socket) {
-        try(socket;
-            BufferedReader reader = new BufferedReader(new InputStreamReader(socket.getInputStream(), UTF_8));
-            PrintWriter writer = new PrintWriter(socket.getOutputStream(), false, UTF_8)) {
-
-            String requestString = requestToString(reader);
-
-            if (requestString.contains("/favicon.ico")) {
-                log("favicon 요청");
-                return;
-            }
-
-            log("http 요청 출력");
-            System.out.println(requestString);
-
-            log("http 응답 생성중");
-            Sleep(1000);
-
-            responseToClient(writer);
-            log("http 응답 생성 끝!");
-
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private String requestToString(BufferedReader reader) throws IOException {
-        StringBuilder sb = new StringBuilder();
-        String line;
-        while((line = reader.readLine()) != null) {
-            if (line.isEmpty()) {
-                break;
-            }
-            sb.append(line).append("\n");
-        }
-        return sb.toString();
-    }
-
-    private void responseToClient(PrintWriter writer) {
-        String body = "<h1>Hello World</h1>";
-        int length = body.getBytes(UTF_8).length;
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("HTTP/1.1 200 OK\r\n");
-        sb.append("Content-Type: text/html; charset=utf-8\r\n");
-        sb.append("Content-Length: ").append(length).append("\r\n");
-        sb.append("\r\n");
-        sb.append(body);
-
-        log("HTTP 응답 정보 출력");
-        System.out.println(sb);
-
-        writer.println(sb);
-        writer.flush();
-    }
-
-    private void Sleep(int i) {
-        try {
-            Thread.sleep(i);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
+            executor.submit(new HttpRequestHandler(socket));
         }
     }
 }
